+++ snack-or-booze/src/SnackDrinkForm.js
@@ -1,70 +1,6 @@
-import React, { useState } from 'react';
 import "./App.css";
-import SnackOrBoozeApi from './Api';
 
-function SnackDrinkForm() {
-  const generateIdFromName = (name) => {
-    return name ? name.toLowerCase().replace(/\s+/g, '-') : '';
-  };
-
-  const [food, setFood] = useState({
-    id: '',
-    type: 'snack',
-    name: '',
-    description: '',
-    recipe: '',
-    serve: ''
-  });
-  console.log(food)
-
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-    setFood({
-      ...food,
-      [name]: value
-    });
-  };
-
-  React.useEffect(() => {
-    const generatedId = generateIdFromName(food.name);
-    setFood(prevState => ({ ...prevState, id: generatedId }));
-  }, [food.name]);
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-    if(food.type === "snack"){
-      try {
-        await SnackOrBoozeApi.addSnack(food)
-        setFood({
-          id: '',
-          type: 'snack',
-          name: '',
-          description: '',
-          recipe: '',
-          serve: ''
-        })
-      } catch (error) {
-        console.error('Error adding snack:', error);
-      }
-    }
-    if(food.type === "drink"){
-      try {
-        await SnackOrBoozeApi.addDrink(food)
-        setFood({
-          id: '',
-          type: 'drink',
-          name: '',
-          description: '',
-          recipe: '',
-          serve: ''
-        })
-        alert('Drink added successfully!');
-      } catch (error) {
-        console.error('Error adding drink:', error);
-      }
-    }
-  }
-
+function SnackDrinkForm({ handleChange, handleSubmit, food }) {
   return (
     <form id="add-food-form">
       <label htmlFor="selectType">Select Type:</label>
